3828: Add posibility to use only localId in getObject request
Opensearch getObject supports two identifiers: pid and id. Pid is the
namespaced data well id and id is the local faust id. It doesn't really
make sense to only use pid in all cases, so here we use the
localIdentifier when localIds is set, instead of before when pid was also used
in this case and constructed with help of agency id.