4060: Ensured that the request object is immutable
The request object is used to generate a cache id by serializing
and hashing the object in "openseach.client.inc" and if the request
object is not immutable this cache id will changes during execution
of the search request. Which will make it impossible to crate the
correct cache id when the search response is parsed that will match
the cache id used for the same request in incoming search request.