4212: Remove object error handling from TingClientRequest
This is a bit misplaced as these errors we are looking for can only
occur in a getObject and not in a search request used by both
collection request and normal search request. Therefore it's better to
handle this error in TingObjectRequest.

If making a collection request for a non-existing ID the search request
will just return nothing and there will be no error, like the ones we
were looking for.

Also, the code didn't event work and failed to cactch any errors anyway,
since it assumed that the object on a collection would be an array, but
this is not the case when making a getObject request where the error
actually has a chance to appear. In getObject request the object on the
collection will NOT be an array but the object itself. As a result of
this, the code was not iterating objects, but instead iterating the
single object failing to ever cactch any errors.