4212: Handle difference in data struture in 5.2
The error element added to objects in opensearch version 5.2+ is
changing the result of the XML parsing: for objects with an
error element the object will not be an array like with "normal"
objects without error element, but instead the object itself. In
earlier versions, where the error is returned as real objects,
the object is still an array. So to be able to handle errors in
all versions, we normalize it here. There seems to be no
difference in the WSDL and XSD, so the new of handling errors
on objects in 5.2+ must be the cause of this change in data
structure. For more info see:

https://platform.dandigbib.org/issues/4212